requirements.txt
# Run 'p2k16-update-requirements' to update this file based on your currently installed set.
#
# This is the specific set of dependencies that is used when installing the application.
# The Ansible setup will install this set.
#
# To update/change dependencies, either update them manually with pip or just rm -rf env/ and run p2k16-run-web again.

attrs==20.2.0
bcrypt==3.1.7
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
cssselect==1.1.0
cssutils==1.0.2
cycler==0.10.0
emails==0.6
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bower==1.3.0
Flask-Env==2.0.0
Flask-Inputs==0.3.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kiwisolver==1.3.2
ldif3==3.2.2
lxml==4.9.1
MarkupSafe==1.1.1
matplotlib==3.4.3
nose==1.3.7
numpy==1.23.1
packaging==21.0
paho-mqtt==1.5.1
pandas==1.3.5
Pillow==9.0.1
pluggy==1.0.0
premailer==3.7.0
psycopg2-binary==2.9.5
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
requests==2.26.0
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Continuum==1.3.11
SQLAlchemy-Utils==0.36.8
stripe==5.5.0
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.6
Werkzeug==1.0.1
WTForms==2.3.3
